backend/requirements.txt
fastapi>=0.104
uvicorn>=0.24
xgboost>=2.0
scikit-learn>=1.3
pandas>=2.1
numpy>=1.26
pyyaml>=6.0
joblib>=1.3
pytest>=8.0
httpx>=0.27
imbalanced-learn>=0.12
torch>=2.0
